% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_Levine32}
\alias{data_Levine32}
\title{Healthy human bone marrow data from Levine, 2015}
\format{
A numeric matrix with 10000 rows and 37 variables.
}
\usage{
data_Levine32
}
\description{
Healthy human bone marrow data from Levine, 2015
}
\keyword{datasets}
