% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSpillMat.R
\name{GetSpillMat}
\alias{GetSpillMat}
\title{Get spillover matrix and cutoffs}
\usage{
GetSpillMat(
  data,
  rows,
  threshold,
  flexrep,
  neighbor,
  quantile,
  seed,
  n_threads
)
}
\arguments{
\item{data}{A numerical matrix.}

\item{rows}{Rows in data to be used for calculating cutoffs.}

\item{threshold}{Spillover max boundary.}

\item{flexrep}{Passed to flexmix. Number of times flexmix is run. See flexmix nrep.}

\item{neighbor}{Abundance spillover channels. Can be 1 or 2.}

\item{quantile}{Quantile used for cutoff calculation.}

\item{seed}{Seed for reproducibility.}

\item{n_threads}{Number of threads used for calculation of cutoffs. Passed to .DeriveCutoffsHelper.}
}
\value{
List of spillover matrix and cutoffs.
}
\description{
Get spillover matrix and cutoffs
}
