% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpillComp.R
\name{SpillComp}
\alias{SpillComp}
\title{Estimate spillover matrix and correct spillover}
\usage{
SpillComp(
  data,
  n = 10000,
  cols = NULL,
  output = NULL,
  threshold = 0.1,
  flexrep = 10,
  neighbor = 1,
  quantile = 0.1,
  seed = 42,
  n_threads = 8
)
}
\arguments{
\item{data}{A numerical matrix.}

\item{n}{Number of sampled events to be used for spillover estimation. n=0 uses all events.}

\item{cols}{Columns to be used for spillover estimation and spillover correction.}

\item{output}{Output FCS filename. Set output = NULL  to not write FCS file.}

\item{threshold}{Spillover max boundary.}

\item{flexrep}{Passed to flexmix. Number of times flexmix is run. See flexmix nrep.}

\item{neighbor}{Abundance spillover channels. Can be 1 or 2.}

\item{quantile}{Quantile used for cutoff calculation. Passed to GetSpillMat.}

\item{seed}{seed for reproducibility.}

\item{n_threads}{Number of threads used for calculation of cutoffs. Passed to GetSpillMat.}
}
\description{
Estimate spillover matrix and correct spillover
}
\references{
Miao, Q, Wang, F, Dou, J, et al. Ab initio spillover compensation in mass cytometry data. Cytometry. 2021; 99: 899– 909. https://doi.org/10.1002/cyto.a.24298
}
